OCSADV-200: make IntegerIsIntegral implicit
